Remove dependancy on host networking driver
The host networking driver only works on Linux hosts, and is not
supported on Docker Desktop for Mac, Docker Desktop for Windows, or
Docker EE for Windows Server.

Host networking, although it provides convinience, it is actually not
needed. Currently, the only service that needs to be exported in the
host is Hugo. For such purposes, manual port publishing can be achived
with "-p".

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>