Allow multiple instances of container
Manually exposing a Hugo server port to the host does not allow a
multiple instances of the container. The problem is that the port is
already allocated on the host when a first instance is run.

This change fixes this issue by using a dynamic allocation of ports on
the host side. This way, each time a container is run, Docker is
responsible for allocating an unused port on a host.

This functionality adds a slight inconvinience to a user as he can not
know which port is allocated on a host since Docker is not activly
reporting this info during running of a container.

Thankfully, Docker provides a command called "ps" which can be used to
check which port on a host is mapped to a port inside of a container.
Example:

$ docker ps
CONTAINER ID   IMAGE                  COMMAND                  CREATED          STATUS          PORTS                     NAMES
3e7789941639   hvarga/tuosde-docker   "entrypoint.sh /usr/…"   8 seconds ago    Up 7 seconds    0.0.0.0:49155->1313/tcp   tender_jackson
d4baf08239cb   hvarga/tuosde-docker   "entrypoint.sh /usr/…"   22 minutes ago   Up 22 minutes   0.0.0.0:49154->1313/tcp   romantic_turing

The output contains a column "PORTS" and from example can be seen that a
port 49155 is allocated on a host which is mapped to a port 1313 in a
container.

Obviously, user can have multiple instances of tuosde-contaner and each
instance will have a unique port allocated on a host. This can be seen
from the above example output. To get the container ID for a specific
container, command "hostname" can be used. This command needs to be
executed from the container shell. Example:

$ hostname
3e7789941639

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>