Fix an issue with prezto during startup
Following error is shown each time a container is started:

/etc/zsh/prezto/init.zsh:unalias:186: no such hash table element: run-help

This is due to a bug in prezto [1].

Problem in tuosde-docker is that each build clones the bleading-edge
version of prezto. The point of a tuosde-docker is to have a
reproducible build and controled development environment. Fetching
newest version of components inside of tuosde-docker on each build can
not provide a stable and controlled environment. Hence, prezto is now
locked-in to a specific version.

[1] https://github.com/sorin-ionescu/prezto/commit/8a967fc1089ce01bc67c7f2540207891c7974230

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>