Use host networking
Podman allow us to use a host networking inside of a container. This
means that container's network stack is not isolated from the host (the
container shares the host's networking namespace), and the container
does not get its own IP address allocated.

This means that a service that exposes its functionality over a TCP port
can now be accessed directly from a host.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>