Add support for running a custom tuosde-docker image
This is the first change in a set of changes that allows a user to
extend tuosde-docker image functionality.

General idea is to have a common or a base tuosde-docker image that can
be extended by anyone to cover needs specific to a project.

There is really no need for each of those images to define its own
"run_tuosde_docker.sh" script since a custom image that inherits from
"hvarga/tuosde-docker" will share the core logic. Hence, this change
adds support for running any custom tuosde-docker images by extending
the common "run_tuosde_docker.sh" script. For a script to be backward
compatible, by default, it will run the core image called
"hvarga/tuosde-docker". User can change the name of the image by
specifing "-i" with the name of an image as an argument to
"run_tuosde_docker.sh".

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>