Add support for extending Neovim configuration
There is no possibility for container images that use tuosde-docker as a
base to extend the default Neovim configuration. While there is always a
possibility to replace the default configuration located in
/usr/share/nvim/sysinit.vim, there was really no mechanism to use the
default one but extend it with project specific Neovim plugins and
custom configuration.

This change adds such mechanism. To extend default Neovim configuration
with project specific one, a child image can install two additional
files into a specific location inside an image, namely:

- /usr/share/nvim/additional_plugins.vim
- /usr/share/nvim/additional_configuration.vim

Both files must contain Vimscript code.

The additional_plugins.vim is a Vimscript file used for installing
additional Neovim plugins. It is executed in the context of a vim-plug
and as such can only contain calls to the Plug command of a vim-plug.

The additional_configuration.vim is a Vimscript file used for extending
Neovim configuration. It is executed at the end of the default Neovim
configuration provided by the base image. It can be used to override
base Neovim configuration and/or add additional configuration specific
to a project.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>