Uplift of major components
Driving idea of this change is to uplift the Neovim to the 0.8 version
to be able to use Lua and modern plugins like Treesitter and Telescope.
Due to this, a huge number of Neovim plugins and other components
received an uplift and reconfiguration.

Result is that Neovim is now more stable, powerful and easier to use.

The Neovim configuration still uses Vimscript which will be ported to
pure Lua configuration in next version which will be a final break to a
compatibility with Vim.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>