Remove hardcoded XAUTHORITY path
By default, xauth will use the file specified by the XAUTHORITY
environment variable or .Xauthority in the user home directory to be
able to authorization connection to the X server.

One problem is that WSL on Windows 10 doesn't have an X server. Although
one can install VcXsrv or Xming, unfortunately this will not set
$XAUTHORITY or create an .Xauthority file. Furthermore, VcXsrv and Xming
don't even require authorisation at all so the .Xauthority file doesn't
even need to exist.

Other problem is that xauth on a GNU/Linux distribution can be
configured to read .Xauthority from anywhere on the system but it MUST
set $XAUTHORITY to point to this location.

Current implementation hardcodes the location of .Xauthority file which
will obviously not work in a multiplatform manner. This change fixes
this by reading the $XAUTHORITY environment variable. If this variable
is not set, it will use $HOME/.Xauthority by default.

Signed-off-by: Hrvoje Varga <hrle@hrle-laptop.home>