Remove entrypoint script
Till 83cc3ca7d1108a10a41b6bbc9b54c3179479effc, the `entrypoint.sh`
script was used to dynamically, during the runtime of a container,
create a user that will have the same username as the one from a host.
Since this information is not baked into an image itself, this is done
on each run of a container. Due to this, a noticable unnecessary delay
was occurring on each run.

Since the conversion from Docker to Podman, this was not needed anymore
due to `--userns=keep-id` argument that can be passed to `podman run`
but the `entrypoint.sh` script is still used to dynamically create a
user home directory.

This change removes a need for a dedicated entrypoint script altogether.
Now, a user with a well know username (developer) is created and used
during a build and the runtime. During the runtime, a host user UID and
GID will be mapped to this developer user into a container. That way,
every change made to a mounted volume from the container will be
properly owned by the host user as was before.

While this change removes a call to an `entrypoint.sh` script, it can be
considered more as a code cleanup long overdue.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>