Complete reconfiguration of a development environment
This change rewrites the whole TUOSDE configuration with the following
ideas in mind:

- Simplicity
- Cohesion
- Maintanance
- Download
- Performance

Container consists of, relativly, a large number of tools each having an
intricate configuration that is extended with all sorts of plugins. Those
tools are smudged together offering a rather clumsy setup.

Due to a lack of cohesion and high complexity, maintaining such
environement has become a burden which lost its practicality. Each
change, no matter how minor it is, whether is just a version bump or
removal of a tool, has a potential to break something. In that regards,
it is becoming more like maintaining a Linux distribution then coupling
together a few different tools.

Such environment has an impact on users as well. It is quite problematic
for a newcomer to get a grasp of what this container provides, how to use
it, and what are the keyboard shorcuts.

The tools are being downloaded from a number of unrelated sources and
repositories. It can be rather difficult if not impossible to build such
image under a security-hardened corporate networks.

Current environment has an somewhat unsatisfactory performance on a big
projects. It can incure a considerable performance impact when navigating
any large project like the Linux or LLVM source code.

This change attempts to fix or at least improve all the aforementioned
painpoints.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>