helix: Disable smart tab
Smart tab is a feature added in Helix 23.12 which is enabled by default
which allows a user in insert mode to jump the cursor instead of actually
inserting a tab character depending on the contents before the cursor.

While the feture has its use cases [1], it does feel unintuitive and the
aim of this commit is to disable it so that pressing tab in insert mode
will always insert a tab character.

[1] https://helix-editor.com/news/release-23-10-highlights/

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>