Switch back to tmux and fzf
While Zellij UI is discoverable as opposed to tmux, tmux feels a bit more
faster and its keybindings don't overlap with other tools. The Zellij
keybindings can be changed but it involves some efforts. There is even an
ongoing discussion [1] to change the default Zellij keybindings.

It seems that maintanance on skim is a bit stagnant as the last change
was more then a year ago. The skim can be also slower then fzf.

Due to above observations, aim of this commit is to replace Zellij and
skim with tmux and fzf, respectively.

[1] https://github.com/zellij-org/zellij/issues/1399
[2] https://github.com/lotabout/skim/issues/317

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>